[MOVE] move position map
