+++ src/plugins/ui_core_views/cell_evaluation/formula_dependency_graph.ts
@@ -1,7 +1,7 @@
 import { positionToZone } from "../../../helpers";
+import { PositionMap } from "../../../helpers/cells/position_map";
 import { recomputeZones } from "../../../helpers/recompute_zones";
 import { CellPosition, UID, Zone } from "../../../types";
-import { PositionMap } from "./position_map";
 import { PositionSet } from "./position_set";
 import { RTreeBoundingBox, RTreeItem, SpreadsheetRTree } from "./r_tree";
