+++ src/plugins/ui_core_views/cell_evaluation/spreading_relation.ts
@@ -1,6 +1,6 @@
 import { positionToZone } from "../../../helpers";
+import { PositionMap } from "../../../helpers/cells/position_map";
 import { CellPosition, UID, Zone } from "../../../types";
-import { PositionMap } from "./position_map";
 import { SpreadsheetRTree } from "./r_tree";
 
 /**
